퀴즈 정답 구현

+++ src/main/java/com/kdt/yts/YouSumback/repository/AnswerOptionRepository.java
@@ -3,5 +3,5 @@ package com.kdt.yts.YouSumback.repository;
 import com.kdt.yts.YouSumback.model.entity.AnswerOption;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnswerOptionRepository extends JpaRepository<AnswerOption, Long> {
+public interface AnswerOptionRepository extends JpaRepository<AnswerOption, Integer> {
 }
